Refactored MSSQL codec implementation
# TDS packet decoder now based on Netty length-field prefixed decoder

# TDS packet encoding centralized in TdsMessageEncoder

This removes duplicated code and will help implement multi packet encoding.

# TDS packet decoding centalized in TdsMessageDecoder

This removes duplicated code and will help maintainers make sense sense of the code as the implementation is now similar to Pg Client.

+++ vertx-mssql-client/src/main/java/io/vertx/mssqlclient/impl/codec/TokenType.java
@@ -14,28 +14,28 @@ package io.vertx.mssqlclient.impl.codec;
 @SuppressWarnings("unused")
 public class TokenType {
 
-  public static final int ALTMETADATA = 0x88;
-  public static final int ALTROW = 0xD3;
-  public static final int COLMETADATA = 0x81;
-  public static final int COLINFO = 0xA5;
-  public static final int DONE = 0xFD;
-  public static final int DONEPROC = 0xFE;
-  public static final int DONEINPROC = 0xFF;
-  public static final int ENVCHANGE = 0xE3;
-  public static final int ERROR = 0xAA;
-  public static final int FEATUREEXTACK = 0xAE;
-  public static final int FEDAUTHINFO = 0xEE;
-  public static final int INFO = 0xAB;
-  public static final int LOGINACK = 0xAD;
-  public static final int NBCROW = 0xD2;
-  public static final int ORDER = 0xA9;
-  public static final int RETURNSTATUS = 0x79;
-  public static final int RETURNVALUE = 0xAC;
-  public static final int ROW = 0xD1;
-  public static final int SESSIONSTATE = 0xE4;
-  public static final int SSPI = 0xED;
-  public static final int TABNAME = 0xA4;
-  public static final int OFFSET = 0x78;
+  public static final short ALTMETADATA = 0x88;
+  public static final short ALTROW = 0xD3;
+  public static final short COLMETADATA = 0x81;
+  public static final short COLINFO = 0xA5;
+  public static final short DONE = 0xFD;
+  public static final short DONEPROC = 0xFE;
+  public static final short DONEINPROC = 0xFF;
+  public static final short ENVCHANGE = 0xE3;
+  public static final short ERROR = 0xAA;
+  public static final short FEATUREEXTACK = 0xAE;
+  public static final short FEDAUTHINFO = 0xEE;
+  public static final short INFO = 0xAB;
+  public static final short LOGINACK = 0xAD;
+  public static final short NBCROW = 0xD2;
+  public static final short ORDER = 0xA9;
+  public static final short RETURNSTATUS = 0x79;
+  public static final short RETURNVALUE = 0xAC;
+  public static final short ROW = 0xD1;
+  public static final short SESSIONSTATE = 0xE4;
+  public static final short SSPI = 0xED;
+  public static final short TABNAME = 0xA4;
+  public static final short OFFSET = 0x78;
 
   private TokenType() {
     // Constants class
